requirements.txt
discord.py>=2.3.0
yt-dlp
python-dotenv
PyNaCl
